Feat: server side loading of plans for subscription page + Loading spinner for the selected plan
